+++ eslint.config.js
@@ -1,5 +1,6 @@
 const base = require('@ton/toolchain');
 const tsEslint = require('@ton/toolchain').tsEslint;
+const globals = require('globals');
 
 module.exports = [
     ...base,
@@ -13,6 +14,15 @@ module.exports = [
             'no-redeclare': 'off',
             '@typescript-eslint/no-redeclare': ['error'],
             '@typescript-eslint/no-explicit-any': 'off',
+        },
+    },
+    {
+        files: ['src/jest/**/*.ts', 'src/jest/**/*.tsx'],
+        languageOptions: {
+            globals: {
+                ...globals.jest,
+                ...globals.node,
+            },
         },
     },
 ];